W-19084772 Remove review order step in one click checkout
