+++ packages/template-retail-react-app/app/pages/checkout-one-click/index.jsx
@@ -5,18 +5,15 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import React, {useEffect, useState} from 'react'
-import {FormattedMessage, useIntl} from 'react-intl'
 import {
     Alert,
     AlertIcon,
     Box,
-    Button,
     Container,
     Grid,
     GridItem,
     Stack
 } from '@salesforce/retail-react-app/app/components/shared/ui'
-import useNavigation from '@salesforce/retail-react-app/app/hooks/use-navigation'
 import {useCheckout} from '@salesforce/retail-react-app/app/pages/checkout-container/util/checkout-context'
 import ContactInfo from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-contact-info'
 import PickupAddress from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-pickup-address'
@@ -25,18 +22,13 @@ import ShippingOptions from '@salesforce/retail-react-app/app/pages/checkout-one
 import Payment from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-payment'
 import OrderSummary from '@salesforce/retail-react-app/app/components/order-summary'
 import {useCurrentBasket} from '@salesforce/retail-react-app/app/hooks/use-current-basket'
-import {useShopperOrdersMutation} from '@salesforce/commerce-sdk-react'
 import {STORE_LOCATOR_IS_ENABLED} from '@salesforce/retail-react-app/app/constants'
 import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
 
 const CheckoutOneClick = () => {
-    const {formatMessage} = useIntl()
-    const navigate = useNavigation()
     const {step} = useCheckout()
-    const [error, setError] = useState()
+    const [error] = useState()
     const {data: basket} = useCurrentBasket()
-    const [isLoading, setIsLoading] = useState(false)
-    const {mutateAsync: createOrder} = useShopperOrdersMutation('createOrder')
     const {passwordless = {}, social = {}} = getConfig().app.login || {}
     const idps = social?.idps
     const isSocialEnabled = !!social?.enabled
@@ -52,24 +44,6 @@ const CheckoutOneClick = () => {
             window.scrollTo({top: 0})
         }
     }, [error, step])
-
-    const submitOrder = async () => {
-        setIsLoading(true)
-        try {
-            const order = await createOrder({
-                body: {basketId: basket.basketId}
-            })
-            navigate(`/checkout/confirmation/${order.orderNo}`)
-        } catch (error) {
-            const message = formatMessage({
-                id: 'checkout.message.generic_error',
-                defaultMessage: 'An unexpected error occurred during checkout.'
-            })
-            setError(message)
-        } finally {
-            setIsLoading(false)
-        }
-    }
 
     return (
         <Box background="gray.50" flex="1">
@@ -97,24 +71,6 @@ const CheckoutOneClick = () => {
                             {isPickupOrder ? <PickupAddress /> : <ShippingAddress />}
                             {!isPickupOrder && <ShippingOptions />}
                             <Payment />
-
-                            {step === 5 && (
-                                <Box pt={3} display={{base: 'none', lg: 'block'}}>
-                                    <Container variant="form">
-                                        <Button
-                                            w="full"
-                                            onClick={submitOrder}
-                                            isLoading={isLoading}
-                                            data-testid="sf-checkout-place-order-btn"
-                                        >
-                                            <FormattedMessage
-                                                defaultMessage="Place Order"
-                                                id="checkout.button.place_order"
-                                            />
-                                        </Button>
-                                    </Container>
-                                </Box>
-                            )}
                         </Stack>
                     </GridItem>
 
@@ -124,43 +80,9 @@ const CheckoutOneClick = () => {
                             showTaxEstimationForm={false}
                             showCartItems={true}
                         />
-
-                        {step === 5 && (
-                            <Box display={{base: 'none', lg: 'block'}} pt={2}>
-                                <Button w="full" onClick={submitOrder} isLoading={isLoading}>
-                                    <FormattedMessage
-                                        defaultMessage="Place Order"
-                                        id="checkout.button.place_order"
-                                    />
-                                </Button>
-                            </Box>
-                        )}
                     </GridItem>
                 </Grid>
             </Container>
-
-            {step === 5 && (
-                <Box
-                    display={{lg: 'none'}}
-                    position="sticky"
-                    bottom="0"
-                    px={4}
-                    pt={6}
-                    pb={11}
-                    background="white"
-                    borderTop="1px solid"
-                    borderColor="gray.100"
-                >
-                    <Container variant="form">
-                        <Button w="full" onClick={submitOrder} isLoading={isLoading}>
-                            <FormattedMessage
-                                defaultMessage="Place Order"
-                                id="checkout.button.place_order"
-                            />
-                        </Button>
-                    </Container>
-                </Box>
-            )}
         </Box>
     )
 }
