+++ packages/template-retail-react-app/app/static/translations/compiled/en-US.json
@@ -907,6 +907,12 @@
       "value": "Remove"
     }
   ],
+  "checkout_payment.button.place_order": [
+    {
+      "type": 0,
+      "value": "Place Order"
+    }
+  ],
   "checkout_payment.button.review_order": [
     {
       "type": 0,
