+++ packages/template-retail-react-app/app/static/translations/compiled/en-XA.json
@@ -1827,6 +1827,20 @@
       "value": "]"
     }
   ],
+  "checkout_payment.button.place_order": [
+    {
+      "type": 0,
+      "value": "["
+    },
+    {
+      "type": 0,
+      "value": "Ƥŀȧȧƈḗḗ Ǿřḓḗḗř"
+    },
+    {
+      "type": 0,
+      "value": "]"
+    }
+  ],
   "checkout_payment.button.review_order": [
     {
       "type": 0,
